requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.0
bcrypt==4.1.2
pdfminer.six==20231228
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.1.4
openai==1.6.1
httpx==0.25.2
Jinja2==3.1.2
huggingface_hub<0.25
itsdangerous==2.1.2
# Notes (comments only; safe for pip):
# Versions pinned for Python 3.9 compatibility and to avoid Gradio schema bool parsing bug.
# If upgrading Python (>=3.10), you can relax pins or move to newer releases.

